+++ data/ddl/package/OSE_PAIEMENT/definition.sql
@@ -16,6 +16,4 @@ CREATE OR REPLACE PACKAGE "OSE_PAIEMENT" AS
 
   FUNCTION get_taux_horaire_date(id_in IN NUMBER, date_val IN DATE) return DATE;
 
-  FUNCTION get_code_taux_remu_legal RETURN VARCHAR2;
-
 END ose_paiement;