+++ module/Intervenant/src/Entity/Db/Statut.php
@@ -12,6 +12,7 @@ use Laminas\Permissions\Acl\Role\RoleInterface;
 use Paiement\Entity\Db\TauxRemu;
 use Plafond\Interfaces\PlafondDataInterface;
 use Plafond\Interfaces\PlafondPerimetreInterface;
+use Service\Entity\Db\TypeVolumeHoraire;
 use UnicaenApp\Service\EntityManagerAwareInterface;
 use UnicaenApp\Service\EntityManagerAwareTrait;
 
@@ -22,8 +23,10 @@ use UnicaenApp\Service\EntityManagerAwareTrait;
  */
 class Statut implements ParametreEntityInterface, RoleInterface, ResourceInterface, EntityManagerAwareInterface, PlafondPerimetreInterface, PlafondDataInterface
 {
-    const CODE_AUTRES       = 'AUTRES';
-    const CODE_NON_AUTORISE = 'NON_AUTORISE';
+    const CODE_AUTRES                      = 'AUTRES';
+    const CODE_NON_AUTORISE                = 'NON_AUTORISE';
+    const ENSEIGNEMENT_MODALITE_SEMESTRIEL = 'semestriel';
+    const ENSEIGNEMENT_MODALITE_CALENDAIRE = 'calendaire';
 
     use ParametreEntityTrait;
     use TypeIntervenantAwareTrait;
@@ -188,13 +191,9 @@ class Statut implements ParametreEntityInterface, RoleInterface, ResourceInterfa
 
     private ?TauxRemu   $tauxRemu                           = null;
 
-    private ?string     $modeServicePrevisionnel            = null;
+    private ?string     $modeEnseignementPrevisionnel       = null;
 
-    private ?string     $modeServiceRealise                 = null;
-
-    private ?string     $modeReferentielPrevisionnel        = null;
-
-    private ?string     $modeReferentielRealise             = null;
+    private ?string     $modeEnseignementRealise            = null;
 
 
 
@@ -1523,66 +1522,63 @@ class Statut implements ParametreEntityInterface, RoleInterface, ResourceInterfa
 
 
 
-    public function getModeServicePrevisionnel(): ?string
+    public function getModeEnseignementPrevisionnel(): ?string
     {
-        return $this->modeServicePrevisionnel;
+        return $this->modeEnseignementPrevisionnel;
     }
 
 
 
-    public function setModeServicePrevisionnel(?string $mode): Statut
+    public function setModeEnseignementPrevisionnel(?string $mode): Statut
     {
-        $this->modeServicePrevisionnel = $mode;
+        $this->modeEnseignementPrevisionnel = $mode;
 
         return $this;
     }
 
 
 
-    public function getModeServiceRealise(): ?string
+    public function getModeEnseignementRealise(): ?string
     {
-        return $this->modeServiceRealise;
+        return $this->modeEnseignementRealise;
     }
 
 
 
-    public function setModeServiceRealise(?string $mode): Statut
+    public function setModeEnseignementRealise(?string $mode): Statut
     {
-        $this->modeServiceRealise = $mode;
+        $this->modeEnseignementRealise = $mode;
 
         return $this;
     }
 
 
 
-    public function getModeReferentielPrevisionnel(): ?string
+    public function isModeEnseignementSemestriel(?TypeVolumeHoraire $typeVolumeHoraire = null): bool
     {
-        return $this->modeReferentielPrevisionnel;
-    }
+        if ($typeVolumeHoraire instanceof TypeVolumeHoraire) {
+            $codeTypeVolumeHoraire = $typeVolumeHoraire->getCode();
+        } else {
+            $codeTypeVolumeHoraire = TypeVolumeHoraire::CODE_PREVU;
+        }
 
+        if ($codeTypeVolumeHoraire == TypeVolumeHoraire::CODE_REALISE) {
+            $modeRealise = $this->getModeEnseignementRealise();
+            if ($modeRealise == self::ENSEIGNEMENT_MODALITE_SEMESTRIEL || is_null($modeRealise)) {
+                return true;
+            }
 
+            return false;
+        } else {
+            $modePrevisionnel = $this->getModeEnseignementPrevisionnel();
+            if ($modePrevisionnel == self::ENSEIGNEMENT_MODALITE_SEMESTRIEL || is_null($modePrevisionnel)) {
+                return true;
+            }
 
-    public function setModeReferentielPrevisionnel(?string $mode): Statut
-    {
-        $this->modeReferentielPrevisionnel = $mode;
+            return false;
+        }
 
-        return $this;
-    }
-
-
-
-    public function getModeReferentielRealise(): ?string
-    {
-        return $this->modeReferentielRealise;
-    }
-
-
-
-    public function setModeReferentielRealise(?string $mode): Statut
-    {
-        $this->modeReferentielRealise = $mode;
-
-        return $this;
+        return true;
     }